cleanup code smells and autoboxing

+++ src/main/java/com/binance/api/client/domain/event/UserDataUpdateEventDeserializer.java
@@ -1,5 +1,7 @@
 package com.binance.api.client.domain.event;
 
+import java.io.IOException;
+
 import com.binance.api.client.domain.event.UserDataUpdateEvent.UserDataUpdateEventType;
 import com.binance.api.client.exception.BinanceApiException;
 import com.fasterxml.jackson.core.JsonParser;
@@ -8,8 +10,6 @@ import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-
-import java.io.IOException;
 
 /**
  * Custom deserializer for a User Data stream event, since the API can return four different responses in this stream.
@@ -32,12 +32,12 @@ public class UserDataUpdateEventDeserializer extends JsonDeserializer<UserDataUp
     String json = node.toString();
 
     final String eventTypeId = node.get("e").asText();
-    final Long eventTime = node.get("E").asLong();
+    final Long eventTime = Long.valueOf(node.get("E").asLong());
     UserDataUpdateEventType userDataUpdateEventType = UserDataUpdateEventType.fromEventTypeId(eventTypeId);
 
     UserDataUpdateEvent userDataUpdateEvent = new UserDataUpdateEvent();
     userDataUpdateEvent.setEventType(userDataUpdateEventType);
-    userDataUpdateEvent.setEventTime(eventTime);
+    userDataUpdateEvent.setEventTime(eventTime.longValue());
 
     if (userDataUpdateEventType == UserDataUpdateEventType.ACCOUNT_UPDATE ||
         userDataUpdateEventType == UserDataUpdateEventType.ACCOUNT_POSITION_UPDATE) {
