+++ src/test/java/com/binance/api/examples/GeneralEndpointsExample.java
@@ -1,5 +1,7 @@
 package com.binance.api.examples;
 
+import java.util.List;
+
 import com.binance.api.client.BinanceApiClientFactory;
 import com.binance.api.client.BinanceApiRestClient;
 import com.binance.api.client.domain.general.Asset;
@@ -7,8 +9,6 @@ import com.binance.api.client.domain.general.ExchangeInfo;
 import com.binance.api.client.domain.general.FilterType;
 import com.binance.api.client.domain.general.SymbolFilter;
 import com.binance.api.client.domain.general.SymbolInfo;
-
-import java.util.List;
 
 /**
  * Examples on how to use the general endpoints.
@@ -23,7 +23,7 @@ public class GeneralEndpointsExample {
     client.ping();
 
     // Check server time
-    long serverTime = client.getServerTime();
+    long serverTime = client.getServerTime().longValue();
     System.out.println(serverTime);
 
     // Exchange info
